requirements.txt
faiss-cpu
sentence-transformers
langchain
streamlit
python-dotenv

pip install -r requirements.txt

pipenv install langchain langchain_community langchain_huggingface faiss-cpu pypdf
pipenv install huggingface_hub
pipenv install streamlit

newenv\Scripts\activate  
streamlit run medibot.py 



